perf(hir): copy a state machine's dispatch into the arm it came from; gate that a benchmark is defined
THREE THINGS, and the second is the one that matters.

1. tailjump (MEASURED M53). A switch inside a loop -- a parser's state machine, a
bytecode interpreter's dispatch -- compiles to ONE dispatch every arm jumps back
to, so every arm re-executes reload/bound-check/index/br. This copies the block
into each predecessor that reaches it by an unconditional jump, so an arm's tail
carries its own copy. Suite 96 against gcc -O2 on Graviton4: EXEC 1.3177 ->
1.3098, INSN 0.9921 -> 1.0073. Duplication, not motion: the copy runs exactly
when that edge is taken, and no instruction is added to or removed from any path.

The census ran BEFORE a line of transform, because M51 had just cost an afternoon
proving the previous obviously-right row fired on nothing: 7 dispatch loops in the
sqlite amalgamation (including sqlite3VdbeExec, 184 arms, header three
instructions), 84 in lua, 29 in the suite, 0 in musl. Both constants come from
that table -- MAX_BLOCK=12 admits every dispatch it found, MAX_GROWTH=800 admits
the largest real case at 184x3=552.

Its ceiling is named rather than hidden. The pass stops at the one block worth
duplicating: m2_http_parse's dispatch loads the byte its arms read and that byte
is used past the arm's first block, so a copy would need a phi and the pass
declines rather than emit a wrong argument list. m2 moves only 2.094 -> 2.075
where M52's counters say 1.95x the instructions are on the table. Two ways past:
SSA reconstruction over the dominance frontier (shared infrastructure -- unroll.rs
names the identical gap), or remat-then-duplicate, which was built and REVERTED
the same hour after the first cut deleted the load without placing it and m2 hung.

2. Four of the ninety-six timed programs were UNDEFINED (MEASURED M54).

tailjump was accused of miscompiling k1_dispatch and k2_live_pressure. Turning it
off left both diverging, which moved the accusation. gcc -O0, gcc -O1 and zcc all
agreed; gcc -O2 was the outlier -- the signature of undefined behaviour, not of a
compiler bug. UBSan: "signed integer overflow: 271182 * 7919 cannot be represented
in type 'int'". Nobody was miscompiling.

Scanning all ninety-six found four: e2_many_args (four of five products left at
int width -- and it is one of the ten worst against -O2, so its 2.41x was never a
measurement), k1_dispatch and k2_live_pressure (overflow twice over), and
u3_shift_var (a rotate guard testing the truth of 32-k, which is 32 when k is 0,
so it shifts by the width). All four fixed; re-scanned 0 of 96.

No gate was at fault. exectime.sh compares zcc against the referee, which is the
right question for a miscompile and is silent about a program neither compiler can
get right; at -O2 it printed DIVERGE for both, correctly. The reader was at fault:
the table was filtered with awk 'NF==5 && $5+0>0', a DIVERGE row carries '-' in
that column, and the geomean was reported as "over 94" without anyone asking where
the other two went. That is the third instance in one session of the same error --
a glob that failed silently and took its command with it, a callgrind job that
held a core for 93 minutes, and now two DIVERGE lines eaten by a filter. The
instrument said so every time.

3. tests/ubscan.sh, in the sci-gate beside provenance. Every timed program must be
defined under UBSan or the gate is RED. Compiled at -O0 on purpose: an optimizer
may delete the undefined operation before the sanitizer sees it, which would make
the gate quieter without making the corpus cleaner. It refuses to pass on an empty
scan, per M39.

Gate 16 PASS / 0 RED at FUZZ_N=300, native on Graviton4 -- ubscan 96 programs all
defined, determinism 187, torture 1694/0, opt-parity 1552/0, csmith 254/0, yarpgen
300/0, musl 479 binaries. cargo 219/0, provenance 31 passes / 52 citations.

+++ tests/bench/suite/e2_many_args.c
@@ -1,3 +1,3 @@
 #include <stdio.h>
-long mix(int a,int b,int c,int d,int e,int f,int g,int h,int i,int j){ return (long)a*j + b*i - c*h + d*g - e*f; }
-int main(void){ long s=0; int k; for(k=0;k<4000000;k++) s += mix(k,k+1,k+2,k+3,k+4,k+5,k+6,k+7,k+8,k+9); printf("%ld\n", s); return 0; }
+long mix(int a,int b,int c,int d,int e,int f,int g,int h,int i,int j){ return (long)a*j + (long)b*i - (long)c*h + (long)d*g - (long)e*f; }
+int main(void){ unsigned long s=0; int k; for(k=0;k<4000000;k++) s += mix(k,k+1,k+2,k+3,k+4,k+5,k+6,k+7,k+8,k+9); printf("%lu\n", s); return 0; }
